+++ common/rx-saga/src/main/kotlin/org/swiften/redux/saga/ReduxSagaHelper.kt
@@ -32,52 +32,12 @@ object ReduxSagaHelper {
     block: suspend CoroutineScope.(P) -> Deferred<R>
   ): ReduxSagaEffect<State, R> = AsyncCallEffect(source, block)
 
-  /** Create a [DelayEffect] */
-  fun <State, R> delay(
-    source: ReduxSagaEffect<State, R>,
-    delayMillis: Long
-  ): ReduxSagaEffect<State, R> = DelayEffect(source, delayMillis)
-
-  /** Create a [MapEffect] */
-  fun <State, P, R> map(
-    param: ReduxSagaEffect<State, P>,
-    block: (P) -> R
-  ): ReduxSagaEffect<State, R> = MapEffect(param, block)
-
-  /** Create a [MapEffect] with [param] */
-  fun <State, P, R> map(param: P, block: (P) -> R) =
-    this.map(this.just<State, P>(param), block)
-
-  /** Create a [CatchErrorEffect] instance. */
-  fun <State, R> catchError(
-    source: ReduxSagaEffect<State, R>,
-    catcher: (Throwable) -> R
-  ): ReduxSagaEffect<State, R> = CatchErrorEffect(source, catcher)
-
-  /** Create a [DoOnValueEffect] instance */
-  fun <State, R> doOnValue(
-    source: ReduxSagaEffect<State, R>,
-    block: (R) -> Unit
-  ): ReduxSagaEffect<State, R> = DoOnValueEffect(source, block)
-
-  /** Create a [FilterEffect] */
-  fun <State, R> filter(
-    source: ReduxSagaEffect<State, R>,
-    selector: (R) -> Boolean
-  ): ReduxSagaEffect<State, R> = FilterEffect(source, selector)
-
   /** Create a [FromEffect] */
   fun <State, R> from(stream: Flowable<R>): ReduxSagaEffect<State, R> =
     FromEffect(stream)
 
   /** Create a [JustEffect] */
   fun <State, R> just(value: R): ReduxSagaEffect<State, R> = JustEffect(value)
-
-  /** Create a [PutEffect] */
-  fun <State, P> put(
-    source: ReduxSagaEffect<State, P>,
-    actionCreator: (P) -> Redux.IAction
-  ): ReduxSagaEffect<State, Any> = PutEffect(source, actionCreator)
 
   /** Create a [SelectEffect] */
   fun <State, R> select(selector: (State) -> R): ReduxSagaEffect<State, R>
@@ -122,11 +82,4 @@ object ReduxSagaHelper {
     { when (it) {is Action -> extract(it); else -> null } },
     block, options
   )
-
-  /** Create a [ThenEffect] on [source1] and [source2] */
-  fun <State, R, R2, R3> then(
-    source1: ReduxSagaEffect<State, R>,
-    source2: ReduxSagaEffect<State, R2>,
-    selector: Function2<R, R2, R3>
-  ) : ReduxSagaEffect<State, R3> = ThenEffect(source1, source2, selector)
 }
