+++ common/saga/src/main/kotlin/org.swiften.redux.saga/CastEffect.kt
