+++ common/saga/src/main/kotlin/org.swiften.redux.saga/FilterEffect.kt
@@ -20,4 +20,4 @@ internal class FilterEffect<State, R>(
 
 /** Invoke a [FilterEffect] on the current [ReduxSagaEffect] */
 fun <State, R> ReduxSagaEffect<State, R>.filter(selector: (R) -> Boolean) =
-  ReduxSagaHelper.filter(this, selector)
+  CommonSagaHelper.filter(this, selector)
