+++ common/saga/src/main/kotlin/org.swiften.redux.saga/MapEffect.kt
@@ -21,4 +21,4 @@ internal class MapEffect<State, P, R>(
 
 /** Invoke a [MapEffect] on the current [ReduxSagaEffect] */
 fun <State, R, R2> ReduxSagaEffect<State, R>.map(block: (R) -> R2) =
-  ReduxSagaHelper.map(this, block)
+  CommonSagaHelper.map(this, block)
