+++ sample/android/build.gradle
@@ -55,6 +55,7 @@ android {
 dependencies {
   implementation fileTree(dir: 'libs', include: ['*.jar'])
   api project(':common:middleware')
+  api project(':common:saga')
   api project(':common:rx-saga')
   api project(':android:ui')
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin"
